requirement.txt
streamlit
playwright
beautifulsoup4
langchain
langchain-core
langchain-groq
chromadb
sentence-transformers
scikit-learn
joblib
python-dotenv
